pyproject.toml
[project]
name = "arkhe"
version = "0.1.0"
description = "Autonomous codebase intelligence — AI-generated maps and dependency graphs for any repository."
requires-python = ">=3.11"

dependencies = [
    "tree-sitter>=0.24.0",
    "tree-sitter-python>=0.24.0",
    "tree-sitter-javascript>=0.23.0",
    "tree-sitter-typescript>=0.23.0",
    "tree-sitter-go>=0.23.0",
    "tree-sitter-rust>=0.23.0",
    "tree-sitter-java>=0.23.0",
    "tree-sitter-ruby>=0.23.0",
    "tiktoken>=0.9.0",
    "python-dotenv>=1.1.0",
    "rich>=13.9.0",
    "pathspec>=0.12.0",
    "groq>=0.13.0",
    "google-genai>=1.0.0",
    "anthropic>=0.49.0",
    "openai>=1.30.0",
    "python-docx>=1.1.0",
    "watchdog>=4.0.0",
    "fastapi>=0.115.0",
    "uvicorn>=0.30.0",
    "jinja2>=3.1.0",
    "python-multipart>=0.0.9",
]

[project.scripts]
arkhe = "main:cli"

[tool.setuptools]
py-modules = ["main"]

[tool.setuptools.packages.find]
where = ["."]

[tool.setuptools.package-data]
templates = ["*.html"]

[tool.pytest.ini_options]
pythonpath = ["."]
asyncio_mode = "auto"

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
]
